+++ Sources/OpenClip/Settings/SettingsCatalog+App.swift
@@ -34,8 +34,6 @@ extension SettingsCatalog {
 
             SettingKey.decisionActiveProvider.erased,
             SettingKey.decisionJevBaseURL.erased,
-            SettingKey.decisionOpenRouterBaseURL.erased,
-            SettingKey.decisionOpenRouterModel.erased,
             SettingKey.decisionLayaCommand.erased,
             SettingKey.decisionToolPresetsJSON.erased,
             SettingKey.decisionLiveAssistDebounceMS.erased